Recognize different bundler environments
When running bundler from Ruby, by default it will inherit the current
bundler environment. For oss-inventory, that isn't what we want.

We need to do two things here:

1. Tell bundler to refresh its environment by wrapping our system calls
in a `Bundler.with_clean_env` block.
2. Ensure that we have local copies of the project's gems (required by
`bundle show`) by doing a `bundle install`.

See
https://stackoverflow.com/questions/25615289/run-bundle-system-command-in-subfolder
for more details on the issue that this addresses.